# docs/api/openapi/api/openapi.yaml
openapi: 3.1.0
info:
  title: Chronos API
  description: "Shift, leave, and sickness planning \u2014 FS Phase 1 implementation."
  version: 0.2.0
paths:
...
  /auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/users:
    get:
      tags:
      - users
      summary: List Users
      operationId: list_users_api_users_get
      parameters:
      - name: team_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Team Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDetail'
                title: Response List Users Api Users Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - users
      summary: Create User
      operationId: create_user_api_users_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/users/{user_id}/export:
    get:
      tags:
      - users
      summary: Export User Data
      operationId: export_user_data_api_users__user_id__export_get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/departments:
    get:
      tags:
      - organization
      summary: List Departments
      operationId: list_departments_api_departments_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DepartmentOut'
                type: array
                title: Response List Departments Api Departments Get
    post:
      tags:
      - organization
      summary: Create Department
      operationId: create_department_api_departments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/departments/{dep_id}:
    delete:
      tags:
      - organization
      summary: Delete Department
      operationId: delete_department_api_departments__dep_id__delete
      parameters:
      - name: dep_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Dep Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/teams:
    get:
      tags:
      - organization
      summary: List Teams
      operationId: list_teams_api_teams_get
      parameters:
      - name: department_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Department Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamOut'
                title: Response List Teams Api Teams Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - organization
      summary: Create Team
      operationId: create_team_api_teams_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/teams/{team_id}:
    delete:
      tags:
      - organization
      summary: Delete Team
      operationId: delete_team_api_teams__team_id__delete
      parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Team Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/delegates:
    get:
      tags:
      - delegates
      summary: List Delegates
      operationId: list_delegates_api_delegates_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DelegateOut'
                type: array
                title: Response List Delegates Api Delegates Get
    post:
      tags:
      - delegates
      summary: Create Delegate
      operationId: create_delegate_api_delegates_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DelegateIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DelegateOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/delegates/{delegate_id}:
    delete:
      tags:
      - delegates
      summary: Delete Delegate
      operationId: delete_delegate_api_delegates__delegate_id__delete
      parameters:
      - name: delegate_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Delegate Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DelegateIn:
      properties:
        principal_id:
          type: string
          format: uuid
          title: Principal Id
        delegate_user_id:
          type: string
          format: uuid
          title: Delegate User Id
        valid_from:
          type: string
          format: date
          title: Valid From
        valid_to:
          type: string
          format: date
          title: Valid To
      type: object
      required:
      - principal_id
      - delegate_user_id
      - valid_from
      - valid_to
      title: DelegateIn
    DelegateOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        principal_id:
          type: string
          format: uuid
          title: Principal Id
        delegate_user_id:
          type: string
          format: uuid
          title: Delegate User Id
        valid_from:
          type: string
          format: date
          title: Valid From
        valid_to:
          type: string
          format: date
          title: Valid To
      type: object
      required:
      - id
      - principal_id
      - delegate_user_id
      - valid_from
      - valid_to
      title: DelegateOut
    DepartmentIn:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
      type: object
      required:
      - name
      title: DepartmentIn
    DepartmentOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
      type: object
      required:
      - id
      - name
      title: DepartmentOut
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
      type: object
      required:
      - username
      - password
      title: LoginRequest
...
    Role:
      type: string
      enum:
      - employee
      - team_lead
      - hr
      - admin
      title: Role
...
    TeamIn:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        department_id:
          type: string
          format: uuid
          title: Department Id
      type: object
      required:
      - name
      - department_id
      title: TeamIn
    TeamOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        department_id:
          type: string
          format: uuid
          title: Department Id
      type: object
      required:
      - id
      - name
      - department_id
      title: TeamOut
    UserCreate:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          minLength: 8
          title: Password
        first_name:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Last Name
        role:
          $ref: '#/components/schemas/Role'
          default: employee
        locale:
          type: string
          title: Locale
          default: en
        time_zone:
          type: string
          title: Time Zone
          default: UTC
        department_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Department Id
        team_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Team Id
      type: object
      required:
      - email
      - password
      title: UserCreate
    UserDetail:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          title: Email
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        role:
          type: string
          title: Role
        locale:
          type: string
          title: Locale
        time_zone:
          type: string
          title: Time Zone
        department_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Department Id
        team_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Team Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - email
      - role
      - locale
      - time_zone
      - created_at
      - updated_at
      title: UserDetail
    UserPublic:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          title: Email
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        role:
          type: string
          title: Role
        locale:
          type: string
          title: Locale
        time_zone:
          type: string
          title: Time Zone
        department_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Department Id
        team_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Team Id
      type: object
      required:
      - id
      - email
      - role
      - locale
      - time_zone
      title: UserPublic
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...